test: add test file for the graph/check implementation
